% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.xword_js.R
\name{as.matrix.xword_js}
\alias{as.matrix.xword_js}
\title{as.matrix.xword_js}
\usage{
\method{as.matrix}{xword_js}(x, ...)
}
\arguments{
\item{x}{data to transformed}

\item{...}{not used, just here for mirroring generic}
}
\description{
as.matrix.xword_js
}
