% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_crossword.R
\name{xword_gen}
\alias{xword_gen}
\title{xword_gen}
\usage{
xword_gen(words, clues = words, tries = 10)
}
\arguments{
\item{words}{character vector of words to put on grid}

\item{clues}{character vector of clues for words}

\item{tries}{number of tries to build script}
}
\description{
xword_gen
}
